dffg

+++ file/edit_data.js
@@ -16,7 +16,7 @@ var sql_insert = function(data, agg_active){
             console.log("1 record inserted");
         }
     });
-    console.logf("check");
+    console.log("check");
     return active_data;
 }
 module.exports = sql_insert;
